Rename utils files to camelCase & delete unnecessary imports
